Add or replace a function in a queue

Functions starting with add_... do add a function to a queue which is processed by the coresponding run_... command.

add(fun, funname, queue)

add_additor(fun)

add_archiver(fun)

add_extractor(fun)

add_pre_processor(fun)

Arguments

fun
function which is run when calling run_...() The functions must not require any arguments!

funname
name of the function

queue
name of queue in getOption("LEEF")

Value

invisiby the function queue.

invisiby the function queue. A list which is processed

invisiby the function queue. A list which is processed

Details

add_additor: Adding a named function to the queue of additors. If the named function already exists will it be replaced.

add_additor: Adding a named function to the queue of additors. If the named function already exists will it be replaced.

add_archiver: Adding a named function to the queue of additors. If the named function already exists will it be replaced.

add_extractor: Adding a named function to the queue of extractors. If the named function already exists will it be replaced.

add_pre_processor: Adding a named function to the queue of pre-processors. If the named function already exists will it be replaced.

Examples

if (FALSE) {
## To add the function `cat` to the `additor` queue
add (fun = cat, .queue = "additor")

## To add the function `paste` to the `extractor` queue
add (cat, "cat", "extractor")
}
add_additor( fun = cat )
add_additor( fun = cat )
add_extractor( fun = paste )
add_pre_processor( fun = paste )