Run the Control Center shiny app included in the LEEF package.

The Control centre app allows the

sanity checks of the raw data

running of the pipeline

control_center(rootdir = ".")

Arguments

rootdir
Directory in which all the data directories can be found.

Value

return value from runApp()

Examples

if (FALSE) {
control_center()
}