Process all ques in the correct order

This function is an example and can be used as a template for processing the queues in a script. It uses the archiver "none" for the raw and pre-processed data, useful for already compressed and large data, e.g. bemovi.

process_raw_comp_none(submitter, timestamp, ...)

Arguments

submitter
name of the submitter of the data to the pipeline. Will be added to the metadata.

timestamp
timestamp of the submission of the data to the pipeline. This should be in the format YYYYMMDD and will be used to identify the sampling day.

...
additional arguments for the different queues

Value

invisibly TRUE

Examples

if (FALSE) {
 process()
}